requirements.txt
celery
flask
flask-cors
flask-babel
flask-migrate
flask-security-too[fsqla,common]
flask-session
flask-socketio
flask-sqlalchemy
gevent-websocket
gunicorn[eventlet]
matplotlib
psycopg2-binary
pytest-cov
python-magic
pyyaml
pytest
redis
requests
simple-websocket
flake8
ipython
supervisor
